move treemap level calculations to seperate module for improved clarity
